bulk: split-and-scatter when SSTBatcher 'fills' range
By keeping track of how many bytes have added to range we just added to,
we can decide to split and scatter a new range when we decide we have
'filled' the one to which we've been adding. When we do so, we split at
the end of what we just added and scatter the next range, i.e. not what
we just imported (since scattering a full range is expensive) but rather
what we can assume we're about to ingest into -- since given that we
just ingested multiple SSTs without hitting a split, it seems likely we
will keep doing so.

With this change backup2tb's 620GB IMPORT of the bank table shows even
growth in ranges across nodes and even leaseholders-per-stores, and runs
in 59m8s or 179.48 MiB/s. That appears to be a large improvement over
the most recent run on master where it took 1h49m24s / 97.02 MiB/s (some
master runs have also been crashing when a node ran out of space).

Release note: none.

+++ pkg/ccl/storageccl/import.go
@@ -192,7 +192,7 @@ func evalImport(ctx context.Context, cArgs batcheval.CommandArgs) (*roachpb.Impo
 		iters = append(iters, iter)
 	}
 
-	batcher, err := bulk.MakeSSTBatcher(ctx, db, MaxImportBatchSize(cArgs.EvalCtx.ClusterSettings()))
+	batcher, err := bulk.MakeSSTBatcher(ctx, db, uint64(MaxImportBatchSize(cArgs.EvalCtx.ClusterSettings())))
 	if err != nil {
 		return nil, err
 	}
